Add support for 2FA TOTP code
If the user has 2FA enabled, they need to pass along with the
credentials a 2FA code from their chosen Time-based one-time (TOTP)
application. In case they don't have it setup, they can just leave
it blank and it will try to authenticate without it.

This code is not stored in the keyring because it is meant to be
disposable and expires fast.